class rearrangment
