+++ src/annbatch/samplers/__init__.py
@@ -1,4 +1,5 @@
-from ._chunk_sampler import ChunkSampler, ChunkSamplerDistributed, ChunkSamplerWithReplacement
+from ._chunk_sampler import ChunkSampler, ChunkSamplerWithReplacement
+from ._chunk_sampler_distributed import ChunkSamplerDistributed
 
 __all__ = [
     "ChunkSampler",
