+++ src/annbatch/samplers/_chunk_sampler.py
@@ -8,44 +8,16 @@ from typing import TYPE_CHECKING
 
 import numpy as np
 
-from annbatch.abc import Sampler
 from annbatch.samplers._utils import get_torch_worker_info
 from annbatch.utils import _spawn_worker_rng, check_lt_1, split_given_size
+
+from ._chunk_sampler_distributed import MaskableSampler
 
 if TYPE_CHECKING:
     from collections.abc import Callable, Iterator
 
     from annbatch.samplers._utils import WorkerInfo
     from annbatch.types import LoadRequest
-
-
-class MaskableSampler(Sampler):
-    """A sampler whose observation range can be restricted via a mask.
-
-    Subclass this to create chunk-based samplers that can be wrapped
-    by :class:`ChunkSamplerDistributed`.
-    """
-
-    _mask: slice
-    _rng: np.random.Generator
-
-    @property
-    def mask(self) -> slice:
-        """The observation range this sampler operates on."""
-        return self._mask
-
-    @mask.setter
-    def mask(self, value: slice) -> None:
-        self._mask = value
-
-    @property
-    def rng(self) -> np.random.Generator:
-        """The random number generator used by this sampler."""
-        return self._rng
-
-    @rng.setter
-    def rng(self, value: np.random.Generator) -> None:
-        self._rng = value
 
 
 class ChunkSamplerBase(MaskableSampler):
